pyproject.toml
[tool.poetry]
name = "sepdpc"
version = "0.1.0"
description = ""
authors = ["Arne Ottens <arne.ottens@starburst.io>","Saravanan Prasadh <saravanan.prasadh@starburstdata.com>"]
readme = "README.md"
packages = [{include = "sepdpc", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
pydantic = "^2.1.1"
pyyaml = "^6.0.1"
deepdiff = "^6.3.1"
python-dotenv = "^1.0.0"
rich = "^13.5.2"
adastra = {git = "https://github.com/ottensa/adastra.git"}

[tool.poetry.scripts]
sepdpc = 'sepdpc.__main__:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
